chore: rename file to kebab-case
